Refactor callPod logic to network.js
